+++ datasets_preparation/prepare_dpo_dataset.py
@@ -5,7 +5,7 @@ import copy
 import time
 
 from functools import partial
-from tokenizer import init_tokenizer
+from tokenization import init_tokenizer
 from datasets import (
     load_dataset,
     interleave_datasets
