+++ datasets_preparation/prepare_instruct_dataset.py
@@ -6,7 +6,7 @@ import copy
 import time
 
 from functools import partial
-from tokenizer import init_tokenizer
+from tokenization import init_tokenizer
 from datasets import (
     load_dataset,
     interleave_datasets
