+++ datasets_preparation/prepare_pretraining_dataset.py
@@ -3,7 +3,7 @@ import numpy as np
 import time
 import copy
 
-from tokenizer import init_tokenizer
+from tokenization import init_tokenizer
 from datasets import (
     load_dataset,
     interleave_datasets
