+++ engine/logging.py
@@ -1,7 +1,7 @@
 import math
 
 from config import TrainingStage, GlobalConfig
-from checkpoints import CheckpointData
+from engine.checkpoints import CheckpointData
 from engine.context import TrainerContext
 from engine.optim import OptimizerPlan
 from engine.core import TrainerState
