+++ tests/conftest.py
@@ -9,7 +9,7 @@ if str(ROOT) not in sys.path:
 import pytest
 import torch
 
-from tokenizer import init_tokenizer
+from tokenization import init_tokenizer
 from models import build_model
 from config import ModelConfig
