+++ tokenization/__init__.py
@@ -1,0 +1,1 @@
+from tokenization.tokenizer import init_tokenizer
