+++ tokenization/tik_tokenizer.py
@@ -1,0 +1,77 @@
+import tiktoken
+
+from tokenization.base import BaseTokenizer
+from pathlib import Path
+from collections import defaultdict
+from tiktoken.load import load_tiktoken_bpe
+
+
+class TikTokenizer(BaseTokenizer):
+    def __init__(self, path, system_prompt):
+        self.special_tokens = defaultdict(int)
+        self.num_reserved_special_tokens = 256
+        self.pat_str = r"(?i:'s|'t|'re|'ve|'m|'ll|'d)|[^\r\n\p{L}\p{N}]?\p{L}+|\p{N}{1,3}| ?[^\s\p{L}\p{N}]+[\r\n]*|\s*[\r\n]+|\s+(?!\S)|\s+"
+
+        self.vocab = load_tiktoken_bpe(path)
+        self.num_base_tokens = len(self.vocab)
+        self.system_prompt = system_prompt
+
+        special_tokens = [
+            '<|begin_of_text|>',
+            '<|end_of_text|>',
+            '<|reserved_special_token_0|>',
+            '<|reserved_special_token_1|>',
+            '<|reserved_special_token_2|>',
+            '<|reserved_special_token_3|>',
+            '<|start_header_id|>',
+            '<|end_header_id|>',
+            '<|reserved_special_token_4|>',
+            '<|eot_id|>',
+        ]
+
+        special_tokens += [
+            f'<|reserved_special_token_{i}|>'
+            for i in range(5, self.num_reserved_special_tokens - 5)
+        ]
+
+        self.special_tokens = {
+            token: self.num_base_tokens + i for i, token in enumerate(special_tokens)
+        }
+
+        self.model = tiktoken.Encoding(
+            name=Path(path).name,
+            pat_str=self.pat_str,
+            mergeable_ranks=self.vocab,
+            special_tokens=self.special_tokens
+        )
+
+        self.vocab_size = self.model.n_vocab
+
+        self.bos_id = self.special_tokens['<|begin_of_text|>']
+        self.eos_id = self.special_tokens['<|end_of_text|>']
+        self.sh_id = self.special_tokens['<|start_header_id|>']
+        self.eh_id = self.special_tokens['<|end_header_id|>']
+        self.eot_id = self.special_tokens['<|eot_id|>']
+        self.pad_id = self.special_tokens['<|reserved_special_token_0|>']
+
+        self.stop_tokens = {
+            self.eos_id,
+            self.eot_id,
+        }
+        
+    def encode(
+        self,
+        text,
+        *,
+        bos=False,
+        eos=False,
+        allowed_special=set(),
+        disallowed_special=()
+    ):
+        tokens = self.model.encode(text, allowed_special=allowed_special, disallowed_special=disallowed_special)
+        if bos: tokens.insert(0, self.bos_id)
+        if eos: tokens.append(self.eos_id)
+        return tokens
+
+    def decode(self, tokens):
+        return self.model.decode(tokens)
